+++ components/updater/card/src/main/java/com/flipperdevices/updater/card/composable/ComposableFirmwareVersionValueChoice.kt
@@ -192,7 +192,7 @@ fun ComposableFirmwareColumn(
     showBackground = true
 )
 @Composable
-fun ComposableUpdaterFirmwareVersionWithChoicePreview() {
+private fun ComposableUpdaterFirmwareVersionWithChoicePreview() {
     FlipperThemeInternal {
         val firmwareVersion = FirmwareVersion(
             channel = FirmwareChannel.RELEASE,