+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 # 1.6.2 - In progress
 
+- [FIX] Fix bug with restoreState navigation
+- [Feature] Add pull to refresh update for update card
+- [Feature] Infrared Emulate (without editing)
+- [Feature] Use as key for fap appcard application id
+- [REFACTOR] Bump deps
+
 # 1.6.1
 
 - [FIX] FapHub build status layout